refactor : 경로 업데이트
