+++ src/pages/Community/components/Comment/Comment.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import type { Comment } from '@/types/post';
-import * as S from './_Comment.style.ts';
+import * as S from './Comment.style.ts';
 import { formatDateTime } from '@/utils/FormatDate';
 import CommentWrite from '../CommentWrite/CommentWrite';
 import KebabMenu from '@/components/common/KebabMenu/KebabMenu';
