+++ src/pages/FindAccount/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import * as S from './_FindAccount.style.ts';
+import * as S from './FindAccount.style.ts';
 import { findEmail, sendPasswordResetCode } from '@/api/auth';
 import { toast } from '@/store/toastStore';
 import CopyIcon from '@/assets/AuthImg/copy.svg';
