+++ src/pages/Mentoring/components/SearchBar/SearchBar.tsx
@@ -1,4 +1,4 @@
-import * as S from './_SearchBar.style.ts';
+import * as S from './SearchBar.style.ts';
 import SearchIcon from '@/assets/mentoringImg/search.png';
 import SearchCancelIcon from '@/assets/mentoringImg/cancel.png';
 import { useSearch } from "@/hooks/useSearch";
