+++ src/pages/Profile/components/WithdrawModal/WithdrawConfirmModal.tsx
@@ -1,4 +1,4 @@
-import * as S from './_WithdrawModal.style.ts';
+import * as S from './WithdrawModal.style.ts';
 import { useNavigate } from 'react-router-dom';
 import { toast } from '@/store/toastStore';
 import { verifyWithdrawalEmailCode } from '@/api/auth';
