+++ src/pages/Profile/components/WithdrawModal/WithdrawModal.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import * as S from './_WithdrawModal.style.ts';
+import * as S from './WithdrawModal.style.ts';
 import WithdrawVerifyModal from './WithdrawVerifyModal';
 
 type Props = {
