+++ src/pages/Profile/components/WithdrawModal/WithdrawVerifyModal.tsx
@@ -1,6 +1,6 @@
 import { useRef, useState, useEffect } from 'react';
 import type { KeyboardEvent } from 'react';
-import * as S from './_WithdrawVerifyModal.style.ts';
+import * as S from './WithdrawVerifyModal.style.ts';
 import arrowSvg from '@/assets/AuthImg/arrow.svg';
 import { sendWithdrawalEmailCode } from '@/api/auth';
 import { toast } from '@/store/toastStore';
