+++ src/pages/Signin/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import * as S from './_Signin.style.ts';
+import * as S from './Signin.style.ts';
 import LogoSvg from '@/assets/AuthImg/AuthLogo.svg';
 import GoogleIcon from '@/assets/Google/Google.svg';
 import { toast } from '@/store/toastStore';
