+++ src/pages/Signup/components/EmailVerifyModal/EmailVerifyModal.tsx
@@ -1,7 +1,7 @@
 import { useRef, useState, useEffect } from 'react';
 import type { KeyboardEvent } from 'react';
 import { useNavigate } from 'react-router-dom';
-import * as S from './_EmailVerifyModal.style.ts';
+import * as S from './EmailVerifyModal.style.ts';
 import arrowSvg from '@/assets/AuthImg/arrow.svg';
 import { verifyEmailCode, sendEmailCode, signup } from '@/api/auth';
 import { toast } from '@/store/toastStore';
